data-access: change namespace name
Change the namespace name of the data access project.
