+++ SensateIoT.Platform.Network.Router/Init/RoutingInitExtensions.cs
@@ -3,8 +3,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using SensateIoT.Platform.Network.DataAccess.Abstract;
-using SensateIoT.Platform.Network.DataAccess.Repositories;
 using SensateIoT.Platform.Router.Common.Caching.Abstract;
 using SensateIoT.Platform.Router.Common.Caching.Routing;
 using SensateIoT.Platform.Router.Common.Collections.Abstract;
@@ -13,6 +11,8 @@ using SensateIoT.Platform.Router.Common.Collections.Remote;
 using SensateIoT.Platform.Router.Common.Services.Processing;
 using SensateIoT.Platform.Router.Common.Settings;
 using SensateIoT.Platform.Router.Data.Abstract;
+using SensateIoT.Platform.Router.DataAccess.Abstract;
+using SensateIoT.Platform.Router.DataAccess.Repositories;
 
 namespace SensateIoT.Platform.Network.Router.Init
 {
