+++ SensateIoT.Platform.Network.Tests/SensateIoT.Platform.Network.Tests.csproj
@@ -18,7 +18,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SensateIoT.Platform.Network.Data\SensateIoT.Platform.Network.Data.csproj" />
     <ProjectReference Include="..\SensateIoT.Platform.Router.Common\SensateIoT.Platform.Router.Common.csproj" />
     <ProjectReference Include="..\SensateIoT.Platform.Router.Data\SensateIoT.Platform.Router.Data.csproj" />
   </ItemGroup>
