+++ SensateIoT.Platform.Router.Common/Collections/Abstract/IInternalRemoteQueue.cs
@@ -7,9 +7,9 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using SensateIoT.Platform.Network.Data.Models;
 using SensateIoT.Platform.Router.Data.Abstract;
 using SensateIoT.Platform.Router.Data.DTO;
+using SensateIoT.Platform.Router.Data.Models;
 
 namespace SensateIoT.Platform.Router.Common.Collections.Abstract
 {
