+++ platform-router.sln
@@ -21,7 +21,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SensateIoT.Platform.Router.Data", "SensateIoT.Platform.Router.Data\SensateIoT.Platform.Router.Data.csproj", "{C6258E1B-4A0F-4D00-B8CA-4A262F42ADB4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SensateIoT.Platform.Network.DataAccess", "SensateIoT.Platform.Network.DataAccess\SensateIoT.Platform.Network.DataAccess.csproj", "{C6A37203-9699-47AC-9EB6-F15B1EBBF01D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SensateIoT.Platform.Router.DataAccess", "SensateIoT.Platform.Router.DataAccess\SensateIoT.Platform.Router.DataAccess.csproj", "{C6A37203-9699-47AC-9EB6-F15B1EBBF01D}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SensateIoT.Platform.Router.Common", "SensateIoT.Platform.Router.Common\SensateIoT.Platform.Router.Common.csproj", "{C6DF3F42-4160-49AD-B095-633573B3A655}"
 EndProject
